+++ config/verbosity.go
@@ -7,40 +7,19 @@ type Verbosity struct {
 	Silent, Quiet, Loud, Debug bool
 }
 
-// LoggingConfiguration produces a logging.Config equivalent to this simpler structure
-func (v Verbosity) LoggingConfiguration() logging.Config {
-	cfg := logging.Config{}
+// UpdateLevel updates the logging level of the provided LogSet
+// to reflect this Verbosity
+func (v Verbosity) UpdateLevel(set *logging.LogSet) {
 	switch {
 	default:
-		cfg.Basic.Level = "warning"
+		// do nothing - if no flag provided, keep configured levels
 	case v.Silent:
-		cfg.Basic.Level = "critical"
-		cfg.Basic.DisableConsole = true
+		set.BeSilent()
 	case v.Quiet:
-		cfg.Basic.Level = "critical"
-	case v.Debug:
-		cfg.Basic.Level = "debug"
+		set.BeQuiet()
 	case v.Loud:
-		cfg.Basic.Level = "extradebug1"
-	}
-
-	return cfg
-}
-
-// LoggingConfigurationToVerbosity is the inverse of
-// Verbosity.LoggingConfiguration.
-func LoggingConfigurationToVerbosity(c logging.Config) *Verbosity {
-	switch c.Basic.Level {
-	default:
-		return &Verbosity{}
-	case "critical":
-		if c.Basic.DisableConsole {
-			return &Verbosity{Silent: true}
-		}
-		return &Verbosity{Quiet: true}
-	case "debug":
-		return &Verbosity{Debug: true}
-	case "extradebug1":
-		return &Verbosity{Loud: true}
+		set.BeChatty()
+	case v.Debug:
+		set.BeHelpful()
 	}
 }